Update omnicoin.php
Everything seems to be working perfectly now! Just updated a few things, and tested it with an insert statement in activate(). The [History] button is only displayed when an address is present now, otherwise it will display "None specified".
We just need to work on the actual mechanism to add the addresses now.